Improved performance, readability and reusability
Passed the process of printing the values to a function and also added a feature that prints the color the returned RGB value represents.
Organized the imports and removed cargos that were not being used.
Now the program tries to remove the screenshot when it first run in order to avoid a behavior where it saved tempscreenshot with a different name due to the file already exist inside the directory.